Feature/rigged animations

+++ interaction_node/detection/detection_algorithm.py
@@ -9,4 +9,5 @@ class DetectionAlgorithim():
       
    def stop_running(self):
          self.isRunning = False
-         self.process.terminate()
+         if self.process != None:
+            self.process.terminate()
